host: Extract RecentCards service
This started as one of the refactorings of container.gts:
move prewarming of recentCards cache to SearchSheet

Tests were failing when I moved it to that component, I
first moved it to OperatorModeState service instead,
but that’s also getting quite large, so I separated
recent cards things out of it.
