+++ packages/host/tests/integration/components/card-delete-test.gts
@@ -15,6 +15,7 @@ import OperatorMode from '@cardstack/host/components/operator-mode/container';
 import type LoaderService from '@cardstack/host/services/loader-service';
 
 import OperatorModeStateService from '@cardstack/host/services/operator-mode-state-service';
+import RecentCardsService from '@cardstack/host/services/recent-cards';
 
 import { CardDef } from 'https://cardstack.com/base/card-api';
 
@@ -624,11 +625,11 @@ module('Integration | card-delete', function (hooks) {
     ];
 
     // creates a recent item
-    let operatorModeStateService = this.owner.lookup(
+    let recentCardsService = this.owner.lookup(
       'service:operator-mode-state-service',
-    ) as OperatorModeStateService;
+    ) as RecentCardsService;
     let mango = await loadCard(`${testRealmURL}Pet/mango`);
-    operatorModeStateService.addRecentCard(mango);
+    recentCardsService.addRecentCard(mango);
 
     await setCardInOperatorModeState([`${testRealmURL}index`]);
     await renderComponent(
